move property names into the filesystem builder directory - they won't beused by any other kind of builder
